changed the firstCubeFromGround heading from null to 270.
Modified the command to move the elevator and wrist down after the robot scoots back from the first scale scoring attempt but before the robot moves to pick up the second cube.

Modified the command to replicate the scooting back and forth behavior for the second scoring in auto from the first scoring in auto.